Delete Cursori
Moved Cursori to its own repository. In theory, it should still be accessible from catsacode.github.io.
A project should be its own repository. It was only put in the website in the first place because I was unsure if it would be be accessible from the GitHub Pages website.